refactoring: - move database management functionality into database package - use new migration framework for database schema migration - trigger migration in database manager constructor - remove support for migration from legacy file based context
