+++ main/adapters/database/database_test.go
@@ -1,25 +1,46 @@
-package repository
+package database
 
 import (
+	"bytes"
 	"context"
 	"encoding/base64"
+	"encoding/json"
+	"fmt"
 	"math/rand"
+	"os"
 	"path/filepath"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/lib/pq"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/ubirch/ubirch-client-go/main/adapters/repository"
+	"github.com/ubirch/ubirch-client-go/main/config"
 	"github.com/ubirch/ubirch-client-go/main/ent"
-	"modernc.org/sqlite"
 )
 
-const testSQLiteDSN = "test.db"
+const (
+	testLoad = 100
+)
 
-func TestDatabaseManager_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+var (
+	testUid          = uuid.MustParse("b8869002-9d19-418a-94b0-83664843396f")
+	testPrivKey      = []byte("-----BEGIN PRIVATE KEY-----\nMHcCAQEEILagfFV70hVPpY1L5pIkWu3mTZisQ1yCmfhKL5vrGQfOoAoGCCqGSM49\nAwEHoUQDQgAEoEOfFKZ2U+r7L3CqCArZ63IyB83zqByp8chT07MeXLBx9WMYsaqn\nb38qXThsEnH7WwSwA/eRKjm9SbR6cve4Mg==\n-----END PRIVATE KEY-----\n")
+	testPubKey       = []byte("-----BEGIN PUBLIC KEY-----\nMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEoEOfFKZ2U+r7L3CqCArZ63IyB83z\nqByp8chT07MeXLBx9WMYsaqnb38qXThsEnH7WwSwA/eRKjm9SbR6cve4Mg==\n-----END PUBLIC KEY-----\n")
+	testSignature, _ = base64.StdEncoding.DecodeString("Uv38ByGCZU8WP18PmmIdcpVmx00QA3xNe7sEB9HixkmBhVrYaB0NhtHpHgAWeTnLZpTSxCKs0gigByk5SH9pmQ==")
+	testAuth         = "650YpEeEBF2H88Z88idG6Q=="
+)
+
+func TestDatabaseManager(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -30,23 +51,23 @@ func TestDatabaseManager_sqlite(t *testing.T) {
 
 	// check not exists
 	_, err = dm.LoadIdentity(testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	_, err = dm.LoadActiveFlag(testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	tx, err := dm.StartTransaction(ctx)
 	require.NoError(t, err)
 	require.NotNil(t, tx)
 
 	_, err = dm.LoadActiveFlagForUpdate(tx, testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	_, err = dm.LoadSignatureForUpdate(tx, testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	_, err = dm.LoadAuthForUpdate(tx, testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	err = tx.Rollback()
 	require.NoError(t, err)
@@ -93,8 +114,13 @@ func TestDatabaseManager_sqlite(t *testing.T) {
 	assert.Equal(t, testIdentity.AuthToken, i.AuthToken)
 }
 
-func TestDatabaseManager_StoreActiveFlag_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_StoreActiveFlag(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -133,8 +159,13 @@ func TestDatabaseManager_StoreActiveFlag_sqlite(t *testing.T) {
 	assert.False(t, active)
 }
 
-func TestDatabaseManager_SetSignature_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_SetSignature(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -180,8 +211,13 @@ func TestDatabaseManager_SetSignature_sqlite(t *testing.T) {
 	assert.Equal(t, newSignature, sig)
 }
 
-func TestDatabaseManager_LoadSignatureForUpdate_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_LoadSignatureForUpdate(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -206,16 +242,27 @@ func TestDatabaseManager_LoadSignatureForUpdate_sqlite(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, tx)
 
-	// try to get lock on database again and wait a second for the lock before context gets canceled
-	ctxWithTimeout, cancelWithTimeout := context.WithTimeout(context.Background(), 200*time.Millisecond)
+	_, err = dm.LoadSignatureForUpdate(tx, testIdentity.Uid)
+	require.NoError(t, err)
+
+	// try to get lock on signature again and wait a second for the lock before context gets canceled
+	ctxWithTimeout, cancelWithTimeout := context.WithTimeout(context.Background(), time.Second)
 	defer cancelWithTimeout()
 
-	_, err = dm.StartTransaction(ctxWithTimeout)
-	assert.EqualError(t, err, "context deadline exceeded")
+	tx2, err := dm.StartTransaction(ctxWithTimeout)
+	require.NoError(t, err)
+
+	_, err = dm.LoadSignatureForUpdate(tx2, testIdentity.Uid)
+	assert.EqualError(t, err, "pq: canceling statement due to user request")
 }
 
-func TestDatabaseManager_StoreAuth_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_StoreAuth(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -261,8 +308,13 @@ func TestDatabaseManager_StoreAuth_sqlite(t *testing.T) {
 	assert.Equal(t, base64.StdEncoding.EncodeToString(newAuth), auth)
 }
 
-func TestDatabaseManager_Ready_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_Ready(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -270,8 +322,27 @@ func TestDatabaseManager_Ready_sqlite(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestDatabaseManager_StoreExisting_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_NotReady(t *testing.T) {
+	t.SkipNow()
+	// this test takes over two minutes before running into a timeout
+	if testing.Short() {
+		t.Skipf("skipping long running test %s in short mode", t.Name())
+	}
+
+	// use DSN that is valid, but not reachable
+	unreachableDSN := "postgres://nousr:nopwd@198.51.100.1:5432/nodatabase"
+
+	_, err := NewDatabaseManager(PostgreSQL, unreachableDSN, 0, false)
+	assert.EqualError(t, err, "dial tcp 198.51.100.1:5432: connect: connection timed out")
+}
+
+func TestDatabaseManager_StoreExisting(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -300,8 +371,13 @@ func TestDatabaseManager_StoreExisting_sqlite(t *testing.T) {
 	assert.Error(t, err)
 }
 
-func TestDatabaseManager_CancelTransaction_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_CancelTransaction(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -322,7 +398,7 @@ func TestDatabaseManager_CancelTransaction_sqlite(t *testing.T) {
 
 	// check transaction was rolled back
 	_, err = dm.LoadIdentity(testIdentity.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	// make sure identity can be stored now
 	ctx, cancel = context.WithCancel(context.Background())
@@ -336,8 +412,13 @@ func TestDatabaseManager_CancelTransaction_sqlite(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestDatabaseManager_StartTransaction_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 1)
+func TestDatabaseManager_StartTransaction(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(2) // since migration is using a database connection, this test needs 2 connections
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -353,60 +434,18 @@ func TestDatabaseManager_StartTransaction_sqlite(t *testing.T) {
 	assert.Nil(t, tx2)
 }
 
-func TestDatabaseManager_StartTransaction2_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
-	require.NoError(t, err)
-	defer cleanUpDB(t, dm)
+func TestDatabaseManager_InvalidTransactionCtx(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
 
-	testIdentity := getTestIdentity()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	tx, err := dm.StartTransaction(ctx)
-	require.NoError(t, err)
-	assert.NotNil(t, tx)
-
-	err = dm.StoreIdentity(tx, testIdentity)
-	require.NoError(t, err)
-
-	err = tx.Commit()
-	require.NoError(t, err)
-
-	tx, err = dm.StartTransaction(ctx)
-	require.NoError(t, err)
-	assert.NotNil(t, tx)
-
-	sig1, err := dm.LoadSignatureForUpdate(tx, testIdentity.Uid)
-	require.NoError(t, err)
-
-	newSig := make([]byte, 64)
-	rand.Read(newSig)
-
-	err = dm.StoreSignature(tx, testIdentity.Uid, newSig)
-	require.NoError(t, err)
-
-	cancel()
-
-	ctx2, cancel2 := context.WithCancel(context.Background())
-	defer cancel2()
-
-	tx2, err := dm.StartTransaction(ctx2)
-	require.NoError(t, err)
-	assert.NotNil(t, tx2)
-
-	sig2, err := dm.LoadSignatureForUpdate(tx2, testIdentity.Uid)
-	require.NoError(t, err)
-	assert.Equal(t, sig1, sig2)
-}
-
-func TestDatabaseManager_InvalidTransactionCtx_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
 	i := ent.Identity{}
-	mockCtx := &mockTx{}
+	mockCtx := &repository.MockTx{}
 
 	err = dm.StoreIdentity(mockCtx, i)
 	assert.EqualError(t, err, "transactionCtx for database manager is not of expected type *sql.Tx")
@@ -430,16 +469,21 @@ func TestDatabaseManager_InvalidTransactionCtx_sqlite(t *testing.T) {
 	assert.EqualError(t, err, "transactionCtx for database manager is not of expected type *sql.Tx")
 }
 
-func TestDatabaseLoad_sqlite(t *testing.T) {
+func TestDatabaseLoad(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
 	wg := &sync.WaitGroup{}
 
-	dm, err := initSQLiteDB(t, 0)
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
 	// generate identities
 	var testIdentities []ent.Identity
-	for i := 0; i < testLoad/10; i++ {
+	for i := 0; i < testLoad; i++ {
 		id := getTestIdentity()
 		id.Uid = uuid.New()
 		testIdentities = append(testIdentities, id)
@@ -463,7 +507,7 @@ func TestDatabaseLoad_sqlite(t *testing.T) {
 	for _, testId := range testIdentities {
 		wg.Add(1)
 		go func(id ent.Identity) {
-			err := checkIdentity(dm, id, dbCheckAuth, wg)
+			err := checkIdentity(dm, id, wg)
 			if err != nil {
 				t.Errorf("%s: %v", id.Uid, err)
 			}
@@ -477,29 +521,50 @@ func TestDatabaseLoad_sqlite(t *testing.T) {
 	//}
 }
 
-func TestDatabaseManager_Retry_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_Retry(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	c, err := getConfig()
+	require.NoError(t, err)
+
+	dm, err := NewDatabaseManager(PostgreSQL, c.DbDSN, 101, true)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
+	wg := &sync.WaitGroup{}
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	tx, err := dm.StartTransaction(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, tx)
+	for i := 0; i < 101; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
 
-	tx2, err := dm.StartTransaction(ctx)
-	require.Error(t, err)
-	require.Nil(t, tx2)
-
-	liteErr, ok := err.(*sqlite.Error)
-	require.True(t, ok)
-	require.Equal(t, 5, liteErr.Code())
+			_, err := dm.StartTransaction(ctx)
+			if err != nil {
+				if pqErr, ok := err.(*pq.Error); ok {
+					switch pqErr.Code {
+					case "55P03", "53300", "53400":
+						return
+					}
+				}
+				t.Error(err)
+			}
+		}()
+	}
+	wg.Wait()
 }
 
-func TestDatabaseManager_StoreExternalIdentity_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_StoreExternalIdentity(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -513,7 +578,7 @@ func TestDatabaseManager_StoreExternalIdentity_sqlite(t *testing.T) {
 	defer cancel()
 
 	_, err = dm.LoadExternalIdentity(ctx, testExtId.Uid)
-	assert.Equal(t, ErrNotExist, err)
+	assert.Equal(t, repository.ErrNotExist, err)
 
 	err = dm.StoreExternalIdentity(ctx, testExtId)
 	require.NoError(t, err)
@@ -535,8 +600,13 @@ func TestDatabaseManager_StoreExternalIdentity_sqlite(t *testing.T) {
 	assert.EqualError(t, err, "context canceled")
 }
 
-func TestDatabaseManager_GetIdentityUUIDs_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_GetIdentityUUIDs(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -562,8 +632,13 @@ func TestDatabaseManager_GetIdentityUUIDs_sqlite(t *testing.T) {
 	}
 }
 
-func TestDatabaseManager_GetExternalIdentityUUIDs_sqlite(t *testing.T) {
-	dm, err := initSQLiteDB(t, 0)
+func TestDatabaseManager_GetExternalIdentityUUIDs(t *testing.T) {
+	// this test communicates with the actual postgres database
+	if testing.Short() {
+		t.Skipf("skipping integration test %s in short mode", t.Name())
+	}
+
+	dm, err := initDB(0)
 	require.NoError(t, err)
 	defer cleanUpDB(t, dm)
 
@@ -591,6 +666,139 @@ func TestDatabaseManager_GetExternalIdentityUUIDs_sqlite(t *testing.T) {
 	}
 }
 
-func initSQLiteDB(t *testing.T, maxConns int) (*DatabaseManager, error) {
-	return NewDatabaseManager(SQLite, filepath.Join(t.TempDir(), testSQLiteDSN), maxConns)
+func getConfig() (*config.Config, error) {
+	configFileName := "config_test.json"
+	fileHandle, err := os.Open(filepath.Join("../../", configFileName))
+	if os.IsNotExist(err) {
+		return nil, fmt.Errorf("%v \n"+
+			"--------------------------------------------------------------------------------\n"+
+			"Please provide a configuration file \"%s\" in the main directory which contains\n"+
+			"a DSN for a postgres database in order to test the database context management.\n\n"+
+			"!!! THIS MUST BE DIFFERENT FROM THE DSN USED FOR THE ACTUAL CONTEXT !!!\n\n"+
+			"{\n\t\"dbDSN\": \"postgres://<username>:<password>@<hostname>:5432/<TEST-database>\"\n}\n"+
+			"--------------------------------------------------------------------------------",
+			err, configFileName)
+	}
+	if err != nil {
+		return nil, err
+	}
+	defer fileHandle.Close()
+
+	c := &config.Config{}
+	err = json.NewDecoder(fileHandle).Decode(c)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(c.DbDSN) == 0 {
+		return nil, fmt.Errorf("missing DSN for test postgres database ('dbDSN') in configuration %s", configFileName)
+	}
+
+	return c, nil
+}
+
+func initDB(maxConns int) (*DatabaseManager, error) {
+	c, err := getConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	return NewDatabaseManager(PostgreSQL, c.DbDSN, maxConns, true)
+}
+
+func cleanUpDB(t assert.TestingT, dm *DatabaseManager) {
+	if dm.driverName == SQLite {
+		time.Sleep(time.Millisecond) // this is here because we are getting SQLITE_BUSY error otherwise
+	}
+
+	err := MigrateDown(dm.db, dm.driverName)
+	assert.NoError(t, err)
+
+	err = dm.Close()
+	assert.NoError(t, err)
+}
+
+func getTestIdentity() ent.Identity {
+	return ent.Identity{
+		Uid:        testUid,
+		PrivateKey: testPrivKey,
+		PublicKey:  testPubKey,
+		Signature:  testSignature,
+		AuthToken:  testAuth,
+	}
+}
+
+func storeIdentity(ctxManager repository.ContextManager, id ent.Identity) error {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tx, err := ctxManager.StartTransaction(ctx)
+	if err != nil {
+		return fmt.Errorf("StartTransaction: %v", err)
+	}
+
+	err = ctxManager.StoreIdentity(tx, id)
+	if err != nil {
+		return fmt.Errorf("StoreIdentity: %v", err)
+	}
+
+	err = tx.Commit()
+	if err != nil {
+		return fmt.Errorf("Commit: %v", err)
+	}
+
+	return nil
+}
+
+func checkIdentity(ctxManager repository.ContextManager, id ent.Identity, wg *sync.WaitGroup) error {
+	defer wg.Done()
+
+	fetchedId, err := ctxManager.LoadIdentity(id.Uid)
+	if err != nil {
+		return fmt.Errorf("LoadIdentity: %v", err)
+	}
+
+	if fetchedId.Uid != id.Uid {
+		return fmt.Errorf("unexpected uuid")
+	}
+
+	if !bytes.Equal(fetchedId.PrivateKey, id.PrivateKey) {
+		return fmt.Errorf("unexpected private key")
+	}
+
+	if !bytes.Equal(fetchedId.PublicKey, id.PublicKey) {
+		return fmt.Errorf("unexpected public key")
+	}
+
+	if !bytes.Equal(fetchedId.Signature, id.Signature) {
+		return fmt.Errorf("unexpected signature")
+	}
+
+	if fetchedId.AuthToken != id.AuthToken {
+		return fmt.Errorf("unexpected auth token")
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tx, err := ctxManager.StartTransaction(ctx)
+	if err != nil {
+		return fmt.Errorf("StartTransaction: %v", err)
+	}
+
+	sig, err := ctxManager.LoadSignatureForUpdate(tx, id.Uid)
+	if err != nil {
+		return fmt.Errorf("LoadSignatureForUpdate: %v", err)
+	}
+
+	err = tx.Commit()
+	if err != nil {
+		return fmt.Errorf("Commit: %v", err)
+	}
+
+	if !bytes.Equal(sig, id.Signature) {
+		return fmt.Errorf("LoadSignatureForUpdate returned unexpected value")
+	}
+
+	return nil
 }
